Heuristics to set memory limit for mutants running with PHPUnit.
Since we know how much memory the initial test suite used,
and only if we know, we can enforce a memory limit upon all
mutation processes. Limit is set to be twice the known amount,
because if we know that a normal test suite used X megabytes,
if a mutants uses a lot more, this is a definite error.

Memory limit is introduced by altering a known temporary php.ini
to include a directive to enable the limit as the very last line.