Use loaded ini settings
The original issue was reported by @senmai in #252:

>right now `php -dmemory_limit=512M` infection has no effect on child processes

The issue is actually not specific to this peculiar parameter but any parameter overriden with the
`-d` option.

Indeed when building the temporary ini file, we are only looking at the settings coming from the
scanned files. Those settings however can be changed with the `-d` option. So provided you have:

```
// php.ini
param=X

// console
$ php -dparam=Y infection
```

We _should_ have in the temporary ini file `param=Y`.