Move bootstrap logic into infection file
Now that bootstrap.php has become much smaller its more logical to
have that logic within the infection entry-point.

We also take advantage of php not un setting variables after a foreach,
we no longer need a define, but instead break out of the loop if we
found the file, and simply check if file is set, and error out otherwise.

We now also include the 'bin' folder in php-cs-fixer and the infection
file as well. Since it would not detect it as a php file otherwise.