Apply max level phpstan to src folder
phpstan.neon has been moved to the devTools folder and it has
been split into two, one for src and one for tests.

Almost all other fixer are type hinting to the interface instead of the
correct class.

We now also throw an error when the phpexectuable finder returns false,
instead of failing on a harder to debug TypeError.