Use our own ConsoleOutput class over SymfonyStyle
Extracted the SymfonyStyle references into a separate class, which makes
the InfectionCommand class no longer in charge of its error messages.

The ConsoleOutput class is in charge of logging anything to the
end user, which keeps other classes clean of this logic.

All other classes should simply call the ConsoleOutput, like so:
` $this->consoleOutput->logInitialTestsDoNotPass(...);`
The ConsoleOutput will then handle the formatting of that message.