Explicitly add the default profile to the list of mutators (#507)
As even if the list is seemingly empty by default, it is not.
If it would actually contain the default profile by default,
it would make the profiles feature more obvious to configure.

Right now it is not immediately obvious from the documentation
that one has to add the default profile to the list, you have
to actually read what is written. With this small change a user
could start adding and changing mutators right away, just after
skimming the example configuration provided.