Include the complete license in headers (#528)
* Include the complete license in headers

Modify PHP-CS-Fixer configuration to include the license verbatim in
headers as required by the terms of the license. Specifically, the license
says that "redistributions of source code must retain the above copyright
notice, this list of conditions and the following disclaimer."

Some licenses like Apache's permit to include only a predefined header,
where this license requires full text to be included to comply with the
license in case of re-distribution, e.g. when using files from a project
in another project. So it is also a matter of courtesy to not to force
any re-distributors to copy and paste the license manually if they wish
to use a file from this project in theirs.

* Actually apply new PHP-CS-Fixer configuraion

* More fixes