Do not round down values in MetricsCalculator
But round them down for display purposes in the relevant logger.

Provided that Infection is being run like so:

    vendor/bin/infection -j2 --min-msi=71.43

It'll report the whole unrounded metric in case of failure:

> The minimum required MSI percentage should be 71.43%, but actual is 71.428571428571%. Improve your tests!

If there isn't a failure, then there's no change in output.

This change is backward compatible because if `floor(71.42) < 72`, then `71.42 < 72` also true.