Move coverage data to the mutation (#733)
This PR has 2 goals

* Reduce the amount of times the same information about coverage is calculated.
* Simplify the API of the `CodeCoverageData` class.

This is done by fetching all the tests a mutation has when it is created. Whether or not it is tested is then determined by checking if the amount of tests is not 0.

Previously, the coverage information was calculated 3 times per mutation.

First it was checked if it had tests, this was then stored in the mutation, along with extra information to later determine what tests covered this mutation.

Then it was calculated twice per line, first to see if it has test for that line, and then to figure out what tests there are for that line. This is now done in one step, during mutation creation.

Because we do this step earlier, we don't have to pass along a bunch of data, simplifying some classes.