Replace `TmpDirCreator` by `TmpDirProvider ` (#878)
With #869 merged, `TmpDirCreator` is now only used for our "production" code, i.e. is no longer used for the tests.

As a result, I'm taking this opportunity to change it: indeed I think the class (and its testing) is overly complex because besides providing a temporary path, it attempts to create it. This also has the unpleasant side effect of potentially creating a temporary directory whenever a `ConfigurationFactory` is used to create a configuration.

In other words: this PR simplifies `TmpDirCreator` by making it responsible only to provide a temporary directory, making the creation of `Configuration` via `ConfigurationFactory` side-effect free.