Massively dumb down the listener, and move logic to factory
The listener was pretty hard to test, since it both created the
objects and called the log method on m.

This is now split up, so the listener just gets an array of
loggers that it should log to after it is finished.

And with some rework in the Logger factory the null logger and the
UnknownLogType became useless and could be removed again.