Rework the traverser implementation (#896)
The goal of this PR is to try to make the traverser we use slightly more robust. I identified to sources of problems with the current implementation:

- We are extending a class instead of relying on the interface, which means we are more likely to be impacted by a change of the parent class
- The parent class API is not adapted to our usage (prioritizing visitors or removing them)

I think there is two good ways to approach the problem:

1. Make use an adapter (the choice of this PR)
2. Use the standard implementation and move the priority related concerns to the traverser factory instead

Regarding 1., since it's the implementation chosen here I will let the code speak for itself.

Regarding 2., I was thinking something along the lines of:

```
final class NodeTraverserFactory
{
    /**
     * @param NodeVisitor[] $extraVisitors
     */
    public function create(array $extraVisitors): PrioritizedVisitorsNodeTraverser
    {
        $traverser = new NodeTraverser();

        $extraVisitors[50] = new NotMutableIgnoreVisitor();
        $extraVisitors[40] = new ParentConnectorVisitor();
        $extraVisitors[30] = new FullyQualifiedClassNameVisitor();
        $extraVisitors[20] = new ReflectionVisitor();

        krsort($extraVisitors, SORT_NUMERIC);

        foreach ($extraVisitors as $visitor) {
            $traverser->addVisitor($visitor);
        }
        
        return $traverser;
    }
}
```

I think however this approach is more fragile since the traverser remains mutable and more error prone (you could add the same visitors again for example it would not fail like it currently [prior and in this PR still] does).