Remove state from the InfectionCommand (#912)
Most elements injected as state for the InfectionCommand are here to make them easier to re-use. However our future goal is to migrate more services which are currently instantiated in the `InfectionCommand` to the `InfectionContainer`. This will ultimately result in that state being removed.

As a result, to make easier this kind of refactoring (moving a service from the command to the container), I'm proposing to remove the state which makes the dependencies easier to follow.