Deprecate mutators that are a semantic addition (#961)
The general consensus is that killing a mutant that is a semantic addition would require a test which tests against unspecified additions. Although it is not excluded to have some value in very specific cases, it is very hard to find reliable case where those mutants performs well and result in the vast majority in false positives.