Introduce IgnoreMutator and a new Mutator interface (#947)
The goal here is to remove `Infection\Mutator\Util\Mutator`. The reason being are that I'm not fond of its API and I do not thing an abstract class is needed here and much less appropriate as an extension point if we ever expose it as such.

The proposal is:

- Move the `Mutator::shouldMutate()` stuff which is mostly about the ignore config in a dedicated class `IgnoreMutator`, relying on decoration instead of inheritance
- Introduce a simpler `Mutator` interface

Since I did not want to introduce too many changes at once, I did keep this class as a BC layer and just marked some elements as `@deprecated` instead.