Mark test cases with `@group integration` when I/O operations are found (#969)
Some test cases executes (or are very likely) I/O operations (read and/or write). Those test cases should be marked as `integration` in order to distinguish them from "pure" test cases as they are more likely to be:

- slower
- flaky if ever run in parallel or sub-process

This does not change much overall, just at _what_ time those tests are executed.

Since it can be quite hard/tedious to detect such cases manually, especially when retro-inspecting the code, I added a test case - which I don't claim to be full-proof, which allows to detect such test cases. And as the consequences of a mistake are minimal, it is a not a big deal if we happen to have a few false-positives.