Remove the MutantInterface (#968)
The `MutantInterface` has only one implementation and since the mutant is an object created internally by infection, it is unlikely to be a part opened up. And even if we do open up that part, then it is likely going to be more appropriate to design the desired interface from the need at that point rather than the existing one.