Introduce `MutatorResolver` (#1006)
The current `MutatorFactory` is doing a bit too much IMO. Indeed it:

- _resolves_ the mutators, i.e. working out from the list of mutator names/class names and profiles with settings, the final list of mutator class names with their settings
- instantiate the mutators from that resolved list

Since so far the instantiation part was relatively trivial, it was ok. But as I'm introducing #1004 and #1005, the instantiate is going to get a bit more interesting. As a result, it makes more sense to split up the factory now with `MutatorResolver` which does the resolving part