Introduce specific configuration objects for the mutators (#1005)
Introduces more specific configuration objects for the mutators to ease the consumption of the mutator settings, effectively removing the need for `Util\MutatorConfig`.

The creation of the config objects as well as further configuration in the mutators (e.g. `MBString` that creates converters based on the config) is all done whilst creating the `Configuration` object. As a result: if there is any error due to a misconfiguration, this will happen early in the process