Introduce a SingletonContainer for tests (#1073)
We have some stateless services used in our test codebase. Since they are used repeatedly, we sometimes try to "cache" them to avoid instantiating them over and over. However this is:

- verbose: requires a static property, static getter (for the right return type)
- error prone & inconsistent: sometimes we use `setUpBeforeClass()` sometimes a getter
- not that efficient: if you have two test cases implementing that strategy you are still instantiating that service 2 times

This PR introduces a `SingletonContainer` to replace `Container` within the tests as well as a few services used for the tests.