Introduce MutantExecutionResult (#1078)
As of now, all our reporting (loggers) are dependent on `MutantProcess` which exposes a lot of mutable state. It also has the downside of requiring a `MutantProcess` for each mutation to work.

I the future, I would like avoid to create the `Mutant` and `MutantProcess` (which are not free) if possible, which does not necessarily exclude the mutation from being reported. 

For this reason, I propose to decouple the reporting from `MutantProcess` by introducing `MutantExecutionResult`. It can be created from a `MutantProcess` for convenience, but it would be easy to create it only from `Mutation` information as well.