Allow to configure ignore globally (#1104)
As of now, if you want to add an `ignore` setting, you need to add it on a mutator basis. This however gets seriously tedious, repetitive and hard to keep in sync if you have a lot of mutators listed.

This PR proposes to allow a global ignore config which will be enriched (not overwritten) by each mutators.