Introduce MutantExecutionResultFactory (#1184)
Ultimately I would like to introduce `ProcessBearer::callback(): (Closure(): void)` which could be called instead of `ParallelProcessRunner::processHandler()`: this allows more granular handling of the "after process" IMO.

This is currently not easily doable since in our case, the callback requires the mutant process itself... One way to circumvent that which also greatly simplifies `MutantProcess` is to move out the detection status operation out of the `MutantProcess`. 

One question one could ask is, does it makes sense? `MutantProcess` is aware of its `Process` and exposes a more friendly API `DetectionStatus` to avoid to have to deal with the `Process` internals (its exit codes) elsewhere. IMO though, `MutantProcess` already exposes its `Process` so having a service doing the interpretation of the process exit codes to deduce the detection status is not a problem.

Another benefit is that it allows `MutantProcess` to revert being a simple (albeit not immutable) VO, i.e. not having any services (the test framework adapter) injected into it