Auto-discovery of the JUnit file (#1193)
One thing that annoys me with the current `--coverage` option is its rigidity. It is cumbersome and unfriendly. Even worse, [our docs on it are wrong](https://infection.github.io/guide/command-line-options.html#coverage) referencing `phpunit.junit.xml` instead of `junit.xml`.

#1107 is an improvement, it helps to:

- fail early if there is a misconfiguration or the coverage does not look good
- provide helpful error messages in case of failure

Although it is an improvement, I think we can go a step further and try to locate automatically the JUnit file and the XML report index file in the provided directory.

Since performances can still be an issue (we would do more I/O lookups), I propose to keep the current expected structure as the default and fallback to a more detailed lookup in case of failure.

This PR is only about the JUnit file. More PRs will be done in order to reach the goal mentioned above.