Make coverage path more tolerant (#1214)
This PR completes the journey started with #1193 and #1194. Now you can provide `--coverage=build/coverage`, it does not matter what path you gave for the XML coverage report or the name for the JUnit report as long as:

- both are in the the given coverage path
- the junit report has a `(.+\.)?junit.xml` pattern

Infection will find the necessary files and figure it out. The previously required path are still suggested and are the "default paths" which are checked first, so there will not be any performance degradation for this feature if you are using the suggested layout