Introduce our own ConsoleLogger (#1263)
The Symfony `ConsoleLogger` is great to be able to make use of the `LoggerInterface` whilst forwarding everything to the output. It however is a bit lacking when it comes to extensibility in regards to:

- easily rewire the log verbosity to a different one
- interpret the context differently for styling options
- is coupled to `OutputInterface`, cannot make use of `SymfonyStyle` or our own `IO`

As a result I initially started with a copy paste of the Symfony one and tweaked it for our needs. Besides introducing this new logger, this PR:

- Replaces the usage of the Symfony one by this new one
- Make `ConsoleOutput` leverage this logger instead of `SymfonyStyle`

There is no behaviour change intended.