Made `logs.badge.branch` capable of accepting a regular expression
By configuring `logs.badge.branch` with a `/`-delimited regular expression, we can collect
reports about a number of branches.

```json
{
    "source": {
        "directories": ["src/"]
    },
    "logs": {
        "badge": {
            "branch": "/^release-.*$/"
        }
    }
}
```

Note that `logs.badge.matchBranchRegex`, initially proposed as an alternative, has also been removed.

We can safely assume that a `branch` value starting and ending with `/` is a regular expression, because
git branches starting or ending with `/` are not valid anyway.

Fixes #1536