Cache phpunit results and run defects first to speed up the Mutation Testing execution
When Infection runs Mutation Analysis for the first time, we can save the information about which Mutant was killed by which test, and then on the subsequent Infection execution (second and more), we can use this cached knowledge and run those test first that killed particular Mutant in a previous call.

This should speed up MT as killed Mutants will be killed from the first "shot".