Treat all function signature nodes as 1-line range to avoid duplicate TestLocations
Previously, `Node` of class `ClassMethod` returned a range for the whole method. Consider the following code:

```
class Test {
    public function findMe() // line 4
    {
        // ...
    } // line 7
}
```

for this code, `ClassMethod` node returned a [4, 7] range instead of [4, 4].

Because of this, later *for each line* (4, 5, 6, 7) we found all the test cases that cover them.

If we have 1 test case `TestClass::test_method_find_me` that covers `findMe()` method, then we returned 4 exactly the same duplicate test cases for `ClassMethod` mutation, one for each of (4, 5, 6, 7) lines.

After this update, only 1 test case is returned for the function signature mutant