Skip analysis when no executed lines with `--git-diff-lines` option
When using the `--git-diff-lines` option, there are situations where a
file can be modified but will not yield any test change, for instance
when only a comment is modified in a PHP source file.

This commit prevents an error being thrown when this very situation
occurs; instead, the analysis is skipped.