Refactor MakefileTest (#1755)
This is built upon #1754 so if you want to check a cleaner diff do skip the first two commits.

This PR is also pure refactoring: there is no new elements being tested, but it is more than simple renaming. The goal here is to prepare for further upcoming changes but I think those changes themselves should be good enough on their own.