refactor(ci): Extract unit & intergation tests from the CI workflow
I think the end-to-end workflow is quite complex and hard to work with.
There is several concerns brought by the nature of end-to-end tests and
they also tend to be inherently slow. This makes it hard to further
tweak the configuration and test a wider range of constraints.

I'm proposing to extract the unit & integration tests. This will allow
to diversify the build to test:

- A wider range of operating systems
- More different PHP versions
- Highest/lowest or locked versions
- Specific Symfony versions

A PoC of that work can be seen in !1786.