fix(ci): Fix bad env configuration (#1793)
I didn't realise the error is so discreet when the configuration file is invalid. I unfortunately didn't pay enough attention to the failing build since almost all my PRs resulted in false-positives recently (and this is something we are actively fixing).

I did a round of review of fixing them all.