Make PHPUnit 10.1 XML coverage report and test cases names with provider compatible with Infection and old format
In PHPUnit >=10 data providers with keys are stored as `Class\\test_method#some key`

In PHPUnit <10 data providers with keys are stored as `Class\\test_method with data set "some key"`.

We need to translate to the old format because this is what PHPUnit <10 and >=10 understand from CLI `--filter` option.

In other words, PHPUnit 10 has a BC break in XML coverage reports that this PR workarounds.