Make `Mutator::getDefinition` return type non-nullable
We are going to extract `Mutator` interface to a separate package, so other developers can require `infection/mutator` and use only that small package with interface(s) to write their custom mutators.

And we 100% must require write definitions for their mutators. If we allow null, developers will simply return null which we are not interested in.

For example Rector does the same - Rector requires providing description / diff / docs for their "rectors". It's a win-win - immediately clear what class is going to do.

The same is for us. Later, even Remedies (3rd parameter will be non-nullable), but this is another story.