Fallback to the simple diff colorizer (previous implementation) for multiline diffs. (#2000)
We introduced inline highlighter for diffs in https://github.com/infection/infection/pull/1974 (thanks to @Slamdunk). However, there was an incorrect assumption that we only have one-line diffs. This is not true.

The case which is currently fails is the following:

```diff
@@ @@
      */
     protected function name()
     {
-        return strtolower(get_class($this));
+        strtolower(get_class($this));
+        return null;
     }
 }
```

I check how `diff-so-fancy` works with multiline diffs, and it turned out that it just skips inline diffs, using simple algorithm, see images below.

So, for the cases when we have multiline diffs, I just added fallback - our old algorithm of colorizing diffs.