Allow setting PHPStan executable custom path
This is needed when phpstan executable is not located in a common vendor/bin/phpstan

For example, in https://github.com/Roave/BetterReflection they have it installed in `./tools`, and Infection can not find it.

----

This implementation is exactly the same as we have for PHPUnit custom path.

Usage:

```json
{
    "$schema": "../../../resources/schema.json",
    "source": {
        "directories": [
            "src"
        ]
    },
    "phpStan": {
        "customPath": "tools/vendor/bin/phpstan"
    }
}
```