Add `--fail-without-result-cache` to PHPStan process (#2182)
As @staab suggested, this will help us identify issues where for some reason there is no result cache, while we expect it to be.

It's ok that PHPStan will fail with exit code = 2, because Ondrej is creating a special error format handler for Infection where if Mutant is killed, PHPStan will return exit code 66.

Exit code 2 will mean mutant is not killed, but PHPStan failed its execution (we will mark it not as `A`, but as for example `E` (error))