Introduce `--id=<mutant-hash>` to run the only one Mutant for killing it (#2226)
* Introduce `--id=<mutant-hash>` to run the only one Mutant for killing it

Related to:

- "Slow feedback loop working with Infection and fixing Mutants for PR" https://github.com/infection/infection/issues/2126
- https://github.com/infection/infection/issues/2191

# Problem

It's not convenient if you want to kill a particular Mutant to run Infection and see the output with N escaped Mutants - it is a noise and you are just wasting the time looking for a needed one.

You need an instant feedback: whether it was killed after updating the tests or whether it's still escaped.

# Solution

Now, we can run

```shell
infection --id=<mutant-hash>
```

and Infection will only run just only produce just only that Mutant.

- when it's killed - you immediately see it
- when it's still escaped - you immediately see a diff in the terminal

Pros:

- we don't wait for `N-1` other mutants tried to be killed

Known issues:

- hash can be invalidated if source code before this Mutant is changed
- hash is non-decodable (well - because it's hash), so we can not extract the file of the Mutant by its hash, to speed up "initial tests suite"

I had an idea to make decodable strings, so that we can encode e.g.

```
src/SomeSourceFile.php:9:16;InstanceOf_
```

so that after decoding we know where (in what file) this Mutant was generated. This would allow us to significantly reduce "initial tests run", because we could under the hood add `--filter=src/SourceFile.php` (related to https://github.com/infection/infection/issues/2191#issuecomment-2973746796). But since this is a completely standalone, very hard problem, for now we can leave without it.

----

how it looks like:

```shell
11) /opt/infection/tests/e2e/Example_Test/src/SourceClass.php:9    [M] Minus [ID] 6fa7eadad5b9fd1a72e4d80c83a61cc3

@@ @@
 {
     public function hello(): string
     {
-        $x = 1 + 2 - 3 * 4;
+        $x = 1 + 2 + 3 * 4;
         return 'hello';
     }
 }
```

* Revert changes in e2e tests

* Move check to MutationTestingRunner.php. Add more tests to kill mutants

* Add mutant id to the TextFileLogger