perf: Optimize the PHPUnit filters (#2407)
In #2415, we are dropping the filters completely in order to fix #1545. However, in practice, there is several strategies we could employ beforehand. My assumption is that although we are spending more time on Infection side to optimise the filter, this should be cheap enough and still bring benefits as I expect the amount of executed tests to be lowered.

# Solution

The idea of this PR is that when we create the filter, we keep track of the number of characters used to build that filter and we set a limit. If that limit is exceeded, instead of immediately bailing out, we try to generate the filters again but employing an optimisation strategy.

The current implementation is done by "level"s, which corresponds to the number of passes made:

- `$optimizationLevel = 0`: no optimisation done whatsoever, this is the default behaviour (and same as on the current main branch).
- `$optimizationLevel = 1`: we drop the provider keys from the tests. Effectively this means we may now execute all the tests from a test method + its provider instead of a subset of it.
- `$optimizationLevel = 2`: we drop the test case name. If it was tests for only one file, there is no data loss. If there is multiple test files being executed and they have methods with the same name, then we now execute all of those methods.
- Beyond this we give up: no filter is applied.

If this is a bit unclear the tests should give clear results.

I also wonder if it would be better to lower the limit more. The current limit is the theoretical limit + a good margin for good measure, but in practice passing too long of a string has a performance impact both because of the shell and the consumer `preg_match()`. So maybe it would be worthwhile to be more aggressive with that limit.


# Performance impact

To execute those tests I:

- Excluded the file `src/TestFramework/PhpUnit/CommandLine/FilterBuilder.php` from the source.
- Excluded the `tests/phpunit/TestFramework/PhpUnit/CommandLine/ArgumentsAndOptionsBuilderTest.php` file from the tests.
- Generate the coverage report.
- Executed infection `$ bin/infection --skip-initial-tests --coverage=...`

| title | result |
|--|--|
| master without `--only-covering-test-cases` | Time: 2m 4s. Memory: 0.18GB. Threads: 10 |
| master with `--only-covering-test-cases` | Time: 2m 3s. Memory: 0.19GB. Threads: 10 |
| this PR without `--only-covering-test-cases` | Time: 1m 58s. Memory: 0.18GB. Threads: 10 |
| this PR with `--only-covering-test-cases` | Time: 1m 45s. Memory: 0.19GB. Threads: 10 |

Obviously this is too poor of a benchmark to be reliable, but at least it doesn't appear there is a severe performance degradation either.