test: Add integration test for coverage trace provider (#2444)
## Goal

The goal of this PR is to bring light about what the final result is, i.e. for a given coverage report, what traces are we getting and what those traces actually are.

This will bring the following benefits:
- Have a clear picture of what `Trace` objects we are consuming look like.
- Ensure we do not have any regressions from reworking any of the underlying code.

## Context

To generate a `Trace`, a lot of works needs to be done. We use about 11 services, wired a special way, in order to navigate through the various coverage reports and construct the `Trace` objects in an efficient way. I believe there is room for improvements there, I have a plan for it, but this is not the scope of this PR.

While doing that work, I was struggling a lot with figuring out how a `Trace` actually looks like in its final form, for a given coverage report. We have a lot of unit tests for the different steps, but none with a clear big picture.

Anyway, this is what brought me here, but I believe this change on its own makes sense even if we do not move forward with [my upstream plan](https://github.com/theofidry/infection/pull/1).


## How it works

To provide real results, a mini-project is shipped for which we can add source code and tests and generate a coverage report. Since coverage reports include absolute path, we cannot commit it. Instead, I went with the solution of copying it and replace the absolute paths by a placeholder and commit this instead. Before executing the tests, we copy the report from the template if it does not exist.

From there we create our service and generate the traces and compare the traces we want.

This test is heavy:
- The expectation list is long: I think it is fine, we can't do much to avoid this and this is meant as a smoke integration test, i.e. not a test we want to update all the time.
- It's a lot of services to instantiate: it is just how it is design-wise at the moment.