test: Introduce a dummy instrumentor for the benchmarks (#2460)
I found myself having to execute those profile scrips a lot for
debugging purposes. However, while debugging, you may not want to
execute Blackfire or you may want to have Xdebug. Commenting the code by
hand each time gets tiring, I think it's better to have a `--debug` flag
to allow to execute the profile without Blackfire.

Note that `DummyInstructor` is currently very simple so it is
questionable to have an interface and dedicated class for this. The
reason I went for this solution is that the `--samples` of Blackfire is
deprecated which means we'll have to do this ourselves in the respective
instructors. So Introducing this class makes the code nicer to read IMO.