test: Introduce a LineReturnNormalizer utility (#2463)
Extract the utility from #2461 and make use of it. It is a simple
utility which IMO makes the code more expressive.

Note the location, I am not entirely sure of the namespace name but I
would like a dedicated directory for generic test utilities without
having them mixed up with source code cases.