refactor: Replace pseudo-enums by enums (#2468)
This PR replaces the pre-PHP 8.1 enums that we had by PHP enums. Along
the way to make things easier a few enum-related utilities have been
introduced.

Some of the enum related utilities were extracted from internal
Webmozarts libraries that I wrote, I plan to move forward to make this
one public, but meanwhile a simplified (the library supports a few more
cases) version. Since it is annoying with the license for PHP-CS-Fixer,
the I copied the license in a place where PHP-CS-Fixer would not replace
it. It's not ideal, but I also don't really want to spend time on this.

There is also a number of escaped mutations, but they are pre-existing to this PR hence fixing them is out of scope.