fix!: Load source files relative to the configuration file location instead of the working directory (#2455)
Fixes #857, #1329.

Documentation PR: https://github.com/infection/site/pull/286.

Currently the paths configured in `source.directories` in
`infection.json5` are relative to the working directory.

On one hand this is awkward because it's been the current behaviour for
so long, some users may rely on it. On the other hand, I think it is
fundamentally fragile, unexpected and last but not least inconsistent
with most (if not all – I may have missed some) other relative paths
that are relative to the configuration file.

Besides the described scenarios of the linked issues, I think the
following scenario (which is the initial one I encountered while working
on #2452) shows how incredibly error-prone this behaviour is. Let's
consider the following structure:

```
infection5.json
src/
  ...
tests/
  benchmark/
    test.php
    infection5.json
    src/
       ...
  
```

Both infection files have the following source:

```json
{
    "source": {
        "directories": [
            "src"
        ]
    },
}

```

And `test.php` looks something like that:

```php
<?php

$container = Container::create()->withValues(
  configFile: __DIR__.'/infection5.json,
  // ...
);
```

Then, if you execute the test from the root directory, the source files
collected will NOT be `/path/to/project/tests/benchmark/src` as you may
expect but `/path/to/project/src`.

This is because in the `SourceFileCollector`, we have:

```php
Finder::create()->in($sourceDirectories)
```

And `$sourceDirectories` is a relative path (technically could be
absolute I guess, but I don't think anyone ever use an absolute one).

In practice this directly affects us for our benchmarks. Whilst I was
trying to get them working (see #2452), I couldn't understand why no
trace was generated.

---------

Co-authored-by: Maks Rafalko <maks.rafalko@gmail.com>