test: Remove dependency of the benchmark sources for end-to-end tests
The end-to-end tests do not execute the benchmarks (neither they
should).

I did see this issue while working on #2452, but left it alone because I
thought of introducing tests of some sorts to ensure we do not break the
benchmarks over time.

This is still in my mind, but I think that will be likely other tests,
i.e. not executed as part of the end-to-end tests. So for now, we can
fix this issue which is a blocker for #2476 and #2452.